+++ runtime/src/bank/tests.rs
@@ -13207,6 +13207,7 @@ fn test_new_for_txn_tests_system_transfer() {
         accounts_lt_hash: AccountsLtHash(LtHash::identity()),
         bank_hash_stats: BankHashStats::default(),
         block_id: None,
+        num_accounts: None,
     };
 
     let bank = Bank::new_for_txn_tests(bank_rc, fields, FeatureSet::all_enabled(), epoch_stakes);
@@ -13384,6 +13385,7 @@ fn test_new_for_block_tests_with_vote_account() {
         accounts_lt_hash: AccountsLtHash(LtHash::identity()),
         bank_hash_stats: BankHashStats::default(),
         block_id: None,
+        num_accounts: None,
     };
 
     let bank = Bank::new_for_block_tests(